Add css sheets to new structure
