--- src/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>Delizia Restaurant</title>
-    <meta name="viewport" content="width=device-width,initial-scale=1" />
-  </head>
-  <body>
-    <div id="content"></div>
-  </body>
-</html>
