+++ src/js/pageSections.js
@@ -1,5 +1,4 @@
-import content from './pageContent';
-import Header from './header';
+import cont from './pageContent';
 
 const Section = (titleContent, bodyContent, type = 'div') => {
   const container = document.createElement('div');
@@ -59,23 +58,63 @@ const Menu = (menuItems) => {
 };
 
 const ContentBox = () => {
-  const cbox = Document.createElement('div');
+  const cbox = document.createElement('div');
   cbox.id = 'content';
   return cbox;
 };
 
+const SectionBox = () => {
+  const sBox = document.createElement('div');
+  sBox.classList.add('section-box');
+  return sBox;
+};
+
+const Header = (logoImg, name, menuLinks) => {
+  const header = document.createElement('header');
+  const brandBox = document.createElement('div');
+  const brandLogo = new Image();
+  const brandName = document.createElement('h1');
+  const navBar = document.createElement('nav');
+  const navMenu = document.createElement('ul');
+
+  Object.keys(menuLinks).forEach(item => {
+    const menuItem = document.createElement('a');
+    const caption = document.createElement('li');
+    caption.innerHTML = item.toString();
+    menuItem.classList.add('nav-menu-item');
+    menuItem.href = menuLinks[item];
+    menuItem.appendChild(caption);
+    navMenu.appendChild(menuItem);
+  });
+
+  brandBox.classList.add('brand-box');
+  brandName.classList.add('brand-name');
+  brandLogo.classList.add('brand-logo');
+  navBar.classList.add('nav');
+  navMenu.classList.add('nav-menu');
+
+  brandName.innerHTML = name;
+  brandLogo.src = logoImg;
+
+
+  brandBox.appendChild(brandLogo);
+  brandBox.appendChild(brandName);
+  navBar.appendChild(navMenu);
+  header.appendChild(brandBox);
+  header.appendChild(navBar);
+
+  return header;
+};
+
 export default function PageSections() {
   const contentBox = ContentBox();
-  const about = Section(content.sections.about.title, content.sections.about.body, 'p');
-  const menu = Section(content.sections.menu.title, Menu(content.sections.menu.menuItems));
-  const contact = Section(content.sections.contact.title, '');
-  const header = Header(
-    content.header.brandLogo,
-    content.header.brandName,
-    content.header.menuOptions,
-  );
+  const sectionBox = SectionBox();
+  const about = Section(cont.sections.about.title, cont.sections.about.body, 'p');
+  const menu = Section(cont.sections.menu.title, Menu(cont.sections.menu.menuItems));
+  const contact = Section(cont.sections.contact.title, '');
+  const header = Header(cont.header.brandLogo, cont.header.brandName, cont.header.menuOptions);
 
   return {
-    header, about, menu, contact, contentBox,
+    header, about, menu, contact, contentBox, sectionBox,
   };
 }
