+++ webpack.config.js
@@ -7,7 +7,7 @@ module.exports = {
   plugins: [
     new CleanWebpackPlugin(),
     new HtmlWebpackPlugin({
-      template: 'src/index.html',
+      title: 'Delizia Restaurant',
     }),
   ],
   output: {
